+++ fl_image_picker/example/app/web/main.dart.js
@@ -45521,7 +45521,7 @@ if(s==null){s=this.c
 s===$&&A.a()}return s}}
 A.eo.prototype={}
 A.jy.prototype={
-E(){return"PickerFromType."+this.b}}
+E(){return"PickerOptionalActions."+this.b}}
 A.GR.prototype={}
 A.Fm.prototype={
 H(a){var s=null,r=new A.bR(B.hj,1,B.R,-1),q=A.GL(B.De,B.hj,s,s,30,s)