+++ format.sh
@@ -1,0 +1,4 @@
+dart format fl_assets_picker/example/lib
+dart format fl_assets_picker/lib
+dart format fl_image_picker/example/lib
+dart format fl_image_picker/lib